+++ frontends/uisvr/src/routes/(public)/signin/+page.svelte
@@ -1,16 +1,15 @@
 <script lang="ts">
-	import Icon from '@iconify/svelte';
 	import { page } from '$app/stores';
+
+	import SigninForm from '$lib/components/forms/SigninForm.svelte';
 
 	import { isFirebaseError } from '$lib/firebase';
 	import type { UserCredential } from '$lib/firebase/auth';
 	import { auth, signInWithEmailAndPassword } from '$lib/firebase/auth';
 
-	let email = '';
-	let password = '';
 	let errorMessage = '';
 
-	const signin = async () => {
+	const signin = async (email: string, password: string) => {
 		let userCredential: UserCredential;
 		try {
 			userCredential = await signInWithEmailAndPassword(auth, email, password);
@@ -30,56 +29,6 @@
 		});
 		window.location.href = $page.url.origin + '/';
 	};
-
-	const signinOnEnter = (e: KeyboardEvent) => {
-		if (e.key === 'Enter') {
-			signin();
-		}
-	};
-	const clearErrorMessage = () => {
-		errorMessage = '';
-	};
 </script>
 
-<div class="prose space-y-4 m-8">
-	<h1 class="mb-4">Sign in</h1>
-
-	{#if errorMessage}
-		<div class="flex">
-			<Icon icon="exclamation-mark-fill" />
-			<p class="text-red-500 ml-2">
-				{errorMessage}
-			</p>
-		</div>
-	{/if}
-
-	<label class="block">
-		<span>Email</span>
-		<input
-			class="mt-1 block w-full"
-			bind:value={email}
-			placeholder="your email address"
-			on:keypress={signinOnEnter}
-			on:change={clearErrorMessage}
-		/>
-	</label>
-
-	<label class="block">
-		<span>Password</span>
-		<input
-			class="mt-1 block w-full"
-			type="password"
-			bind:value={password}
-			placeholder="your password"
-			on:keypress={signinOnEnter}
-			on:change={clearErrorMessage}
-		/>
-	</label>
-	<div class="flex">
-		<button class="btn btn-primary flex-none mt-4 h-12" on:click={signin}>Sign in</button>
-		<div>
-			<a class="btn btn-neutral mt-4 ml-8 h-12" color="alternative" href="/signup">Sign up</a>
-			<div class="ml-8">Sign up if you don't have an account.</div>
-		</div>
-	</div>
-</div>
+<SigninForm bind:errorMessage onSignin={signin} />
